# bsn-fe/src/app/services/models/borrowed-book-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BorrowedBookResponse {
    authorName?: string;
    id?: number;
    isbn?: string;
    rate?: number;
    returnApproved?: boolean;
    returned?: boolean;
    title?: string;
  }